Amend mac assembly script  to --skip-notary
For people who don't have a developer account, attempting to notarize doesn't work. I propose adding this optional flag `--skip-notary` to mac `assemble.sh` script which skips the notary process but still signs for local use. I've adjusted the build instructions to specify this. If the flag isn't in `$2` it will work as normal.